Refactor project structure
- Refactored import statements in various modules to point to the new chalice_support package
