+++ agave/chalice_support/__init__.py
@@ -1,3 +1,2 @@
 __all__ = ['RestApiBlueprint']
-
 from .rest_api import RestApiBlueprint
