+++ examples/chalicelib/models/files.py
@@ -1,7 +1,7 @@
 from mongoengine import Document, StringField
 
-from agave.models import BaseModel
-from agave.models.helpers import uuid_field
+from agave.chalice_support.models import BaseModel
+from agave.chalice_support.models.helpers import uuid_field
 
 
 class File(BaseModel, Document):
