+++ examples/chalicelib/models/transactions.py
@@ -1,7 +1,7 @@
 from mongoengine import Document, FloatField, StringField
 
-from agave.models import BaseModel
-from agave.models.helpers import uuid_field
+from agave.chalice_support.models import BaseModel
+from agave.chalice_support.models.helpers import uuid_field
 
 
 class Transaction(BaseModel, Document):
