+++ examples/chalicelib/resources/accounts.py
@@ -2,7 +2,7 @@ import datetime as dt
 
 from chalice import Response
 
-from agave.filters import generic_query
+from agave.core.filters import generic_query
 
 from ..models import Account as AccountModel
 from ..validators import AccountQuery, AccountRequest, AccountUpdateRequest
