+++ examples/chalicelib/resources/transactions.py
@@ -1,4 +1,4 @@
-from agave.filters import generic_query
+from agave.core.filters import generic_query
 
 from ..models.transactions import Transaction as TransactionModel
 from ..validators import TransactionQuery
